+++ src/components/pages/Tokens.jsx
@@ -1,9 +1,11 @@
+import moment from 'moment'
 import React, { useContext, useEffect, useReducer } from 'react'
 
 import { Main } from 'components/templates/Main'
-import { Tokens as TokensOrganism } from 'components/organisms/Tokens'
 import { parseJwt } from 'helpers/jwt'
 import { ApiContext } from 'providers/ApiProvider'
+
+import classNames from './Tokens.scss'
 
 export const Tokens = () => {
   const [api, isBusy] = useContext(ApiContext)
@@ -22,14 +24,17 @@ export const Tokens = () => {
 
   return (
     <Main>
-      <TokensOrganism tokens={tokens} onCreateToken={onCreateToken} onDelete={onDelete} createDisabled={isBusy} />
+      <section className={classNames.tokens}>
+        <h1>API Tokens</h1>
+        <h2>Manage your API Tokens by authenticating with the Frost API</h2>
+        <TokenTable tokens={tokens} onDelete={onDelete} />
+        <button className={classNames.create} onClick={onCreateToken} disabled={isBusy}>Create API Token</button>
+      </section>
     </Main>
   )
 }
 
 const tokenReducer = (tokens, { type, payload }) => {
-  console.log('tokenReducer', type, payload)
-
   if (payload === undefined)
     return tokens
 
@@ -62,3 +67,42 @@ const serializedTokenToToken = serializedToken => ({
   ...parseJwt(serializedToken),
   serializedToken,
 })
+
+const TokenTable = ({ tokens, onDelete }) => (
+  <table>
+    <thead>
+    <tr>
+      <td>#</td>
+      <td>Token</td>
+      <td>Creation</td>
+      <td>Expiry</td>
+      <td>Actions</td>
+    </tr>
+    </thead>
+    <tbody>
+    {
+      tokens?.map((token, index) =>
+        <TokenTableRow token={token} key={token.serializedToken} index={index + 1} onDelete={() => onDelete(token.serializedToken)} />
+      )
+    }
+    </tbody>
+  </table>
+)
+
+const TokenTableRow = ({ token, index, onDelete }) => {
+  const firstAndLastCharacters = (s, n) => s.slice(0, n) + '...' + s.slice(-n)
+
+  const formattedIat = moment.unix(token.iat).format('L')
+  const formattedIatISO = moment.unix(token.iat).format()
+  const formattedSerializedToken = firstAndLastCharacters(token.serializedToken, 20)
+
+  return (
+    <tr>
+      <td>{index}</td>
+      <td>{formattedSerializedToken}</td>
+      <td title={formattedIatISO}>{formattedIat}</td>
+      <td>Never</td>
+      <td><button className={classNames.delete} onClick={onDelete}>Remove</button></td>
+    </tr>
+  )
+}
